% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_col_pvalues.R
\name{scale_col_pvalues}
\alias{scale_col_pvalues}
\title{scale_col_pvalues is an alias of scale_color_pvalues}
\usage{
scale_col_pvalues(palette = "red")
}
\arguments{
\item{palette}{The colors palette, allowed values: red, blue, green}
}
\value{
ggplot object
}
\description{
scale_col_pvalues is an alias of scale_color_pvalues
}
