% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_fu.R
\name{theme_fu}
\alias{theme_fu}
\title{theme fu}
\usage{
theme_fu()
}
\value{
ggplot object
}
\description{
This is an example custom theme for an ggplot object.
It has been evolved from here https://rstudio-pubs-static.s3.amazonaws.com/570580_1ee08ab4cad0497fa5718f12bbac6f5c.html
}
